requirements.txt
# Flask==3.0.0
# dash
# local
# dash==2.14.2
# pip install dash==2.5.0
pandas
gunicorn
python-dotenv

Flask==2.2
Werkzeug==2.2.2
# dash==2.5.0

# https://community.plotly.com/t/multipage-app-with-dash-2-5-0/68469
# fix multi page bug can register only 1 page
dash==2.6.1
dash-auth

langchain
langchain-community
langchain-openai

SQLAlchemy
sqlmodel
psycopg2-binary~=2.9.3
names
python-dotenv
dash_bootstrap_templates